% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{benders_check}
\alias{benders_check}
\title{Check that the ANTARES study is ready for benders decomposition}
\usage{
benders_check(benders_options, opts)
}
\arguments{
\item{benders_options}{list of benders decomposition options, as returned by
\code{\link{read_options}}.}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
boolean, indicating whether or not the test are fullfilled
}
\description{
Check that the ANTARES study is ready for benders decomposition
}

