% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_output_benders}
\alias{clean_output_benders}
\title{Delete the output of the benders iterations which are not usefull anymore}
\usage{
clean_output_benders(best_solution, unique_key,
  opts = antaresRead::simOptions())
}
\arguments{
\item{best_solution}{identifier of the best solution so far 
(the output related to this iteration will not be removed)}

\item{unique_key}{unique string to identify the output of the current expansion problem}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{

}
\description{
Delete the output of the benders iterations which are not usefull anymore
}
