% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benders_tools.R
\name{convergence}
\alias{convergence}
\title{Check the convergence of the expansion problem}
\usage{
convergence(best_sol, best_under, benders_options)
}
\arguments{
\item{best_sol}{cost of best solution}

\item{best_under}{cost of best under estimator}

\item{benders_options}{list of benders decomposition options, as returned by
\code{\link{read_options}}.}
}
\value{
logical, indicating weither or not the expansion problem has converged within the optimality gap
}
\description{
Check the convergence of the expansion problem
}

