% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_selection.R
\name{convert}
\alias{convert}
\title{Convert to_simulate file into independent mc_years and weeks list}
\usage{
convert(to_simulate, mc_years, weeks)
}
\arguments{
\item{to_simulate}{data.frame with week, year and weight column}

\item{mc_years}{all mc_years of the study}

\item{weeks}{all weeks of the study}
}
\value{
return a list of output containing the playlist of mc years (out$mc_years)
and the playlist of weeks (out$weeks)
}
\description{
to_simulate contains pair (mc_year, week) with their weight
the weight is between zero and one :
   - a null weight implies that the pair (mc_year, week) shouldn't be
     simulated in the next benders iteration
   - a weight equal to 1 implies that the pair is particularly relevant
     ans should be simulated in the next benders iteration
     
this function therefore choses a mc_years list and a weeks list which  
covers most of the weeks with strictly positive weight, but not too many 
weeks with a weight equal to zero
}

