% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{default_report_file}
\alias{default_report_file}
\title{Return the default file name for the report}
\usage{
default_report_file(opts = simOptions())
}
\arguments{
\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
name of the first area that the link connects
}
\description{
Return the default file name for the report
}

