% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{delete_antares_output}
\alias{delete_antares_output}
\title{Delete the output of an ANTARES study}
\usage{
delete_antares_output(simulation_name, opts = antaresRead::simOptions())
}
\arguments{
\item{simulation_name}{name of the simulation whose output will be deleted}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{

}
\description{
Delete the output of an ANTARES study
}
