% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{enable_custom_filtering}
\alias{enable_custom_filtering}
\title{Enable or disable the custom filtering of the output}
\usage{
enable_custom_filtering(enable = TRUE, opts = antaresRead::simOptions())
}
\arguments{
\item{enable}{Should the custom filtering be enabled (\code{TRUE}) or disabled (\code{FALSE}) ?}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{enable_custom_filtering} is a function which modifies the input file of an ANTARES
study by enabling (or disabling) the custom filtering of the output
}
