% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.r
\name{enable_hurdle_costs}
\alias{enable_hurdle_costs}
\title{Enable or disable the hurdle cost of a link or a list of links}
\usage{
enable_hurdle_costs(link_names, enable = TRUE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{link_names}{Name of one link, or vector with several link names}

\item{enable}{Should the hurdle costs be enabled (\code{TRUE}) or disabled (\code{FALSE}) ?}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{enable_hurdle_costs} is a function which modifies the input file of an ANTARES
study and enable (or disable) the hurdle costs of a link
}
