% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{enable_year_by_year}
\alias{enable_year_by_year}
\title{Enable or disable the year by year output writing}
\usage{
enable_year_by_year(enable = TRUE, opts = antaresRead::simOptions())
}
\arguments{
\item{enable}{Should the year by year outputs be written (\code{TRUE}) or not (\code{FALSE}) ?}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{enable_year_by_year} is a function which modifies the input file of an ANTARES
study by activating (or disactivating) the year by year writing of the output
}
