% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{filter_output_links}
\alias{filter_output_links}
\title{Modify the filters on the output of the links}
\usage{
filter_output_links(links, filter, type, opts = simOptions())
}
\arguments{
\item{links}{Link, or vector of links whose filter will be changed.}

\item{filter}{Vector of filters to adopt for the area. Can contain "annual", "monthly", "weekly", 
"daily", "hourly" or a combination of them.}

\item{type}{On which type of output the filters will applied ? The MC synthesis (\code{type = "synthesis"}) ? The 
year by year outputs (\code{type = "year-by-year"}) ? Or both(\code{type = c("synthesis", "year-by-year")}) 
?}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{filter_output_links} is a function which modifies the input file of an ANTARES
study by updating the output filters of some links
}

