% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.r
\name{from}
\alias{from}
\title{Return the first area that a link connects}
\usage{
from(link_name)
}
\arguments{
\item{link_name}{Name of one link}
}
\value{
name of the first area that the link connects
}
\description{
Return the first area that a link connects
}
