% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capacity_profile.R
\name{get_capacity_profile}
\alias{get_capacity_profile}
\title{Build capacity time series}
\usage{
get_capacity_profile(installed_capacity, load_factor_profile)
}
\arguments{
\item{installed_capacity}{new installed capacity}

\item{load_factor_profile}{Time series of scalar of actual percentage of available capacity}
}
\value{
capacity time series
}
\description{
Given the installed capacity and the load factor profile,
return the new capacity profile.
The capacity profile might not be equal to a simple multiplication between the
installed capacity and the load factor as it needs to be rounded to match with 
the ANTARES data format. The rounding is made so as to respect the total producible
energy of all the year.
}

