% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_defaillance.R
\name{get_margins}
\alias{get_margins}
\title{Compute generation margin in a given area on an Antares Study.}
\usage{
get_margins(area = "fr", cluster_name = "gas_pcomp_peak", LOLD = 3,
  tolerance = 0.01, unit_size = 100, abaque = function(c) {    
  return(-3627 * log(c) + 3723.4) }, initial_margin = 0, path_solver,
  display = TRUE, clean = TRUE, parallel = TRUE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{area}{Area in which the margin is computed}

\item{cluster_name}{Name of the complementary cluster if the margin is negative (lack of production).
The cluster can already exist, otherwise it will be created.
(example : If cluster_name = "gas_pcomp_peak" and area = "fr", the final name of the cluster will be "fr_gas_pcomp_peak")}

\item{LOLD}{Adequacy criteria of the area in terms of Loss of Load Duration (given in h/year)}

\item{tolerance}{Algorithmic tolerance about the adequacy criteria. Used as stop criterion. 
(example : if tolerance = 0.5 and LOLD = 3.0, the aimed range of LOLD is between 2h30 and 3h30)}

\item{unit_size}{Minimal step of the generation margin. Given in MW.}

\item{abaque}{Function abaque : marge = f(LOLE)}

\item{initial_margin}{First margin tested in the function. Given in MW.}

\item{path_solver}{Character containing the Antares Solver path}

\item{display}{Logical. If \code{TRUE} the advancement of the function
if displayed in the console}

\item{clean}{Logical. If \code{TRUE} the output of the ANTARES simulations run by the
package will be deleted (except for the output of the simulation which brings
to the best solution).}

\item{parallel}{Logical. If \code{TRUE} the ANTARES simulations will be run in parallel mode (Work
only with ANTARES v6.0.0 or more). In that case, the number of cores used by the simulation
is the one set in advanced_settings/simulation_cores (see ANTARES interface).}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{

}
\description{
- A positive generation margin means that the generation fleet of the area is over-sized 
  and the LOLD (Loss Of Load Duration) is below the adequacy criteria (typically 3h/y). In that
  case, the margin corresponds to the extra constant load that the generation fleet can still supply while
  respecting strictly the adequacy criteria.
  
  - A negative generation margin means that the generation fleet is under-sized and the LOLD
  is above the adequacy criteria. In that case, the margin corresponds to the extra "perfect" production
  capacity required to meet the adequacy criteria. A "perfect" capacity is a capacity which is available
  all the time (i.e. without maintenance/outages).
}
