% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{get_playlist}
\alias{get_playlist}
\title{Get playlist of simulated MC years}
\usage{
get_playlist(opts = antaresRead::simOptions())
}
\arguments{
\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Returns a vector of the identifier of the simulated MC year
}
\description{
\code{get_playlist} gives the identifier of the MC years which
will be simulated, taking into account the potential use of a 
playlist which can disactivate some MC years
}

