% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sensitivity.R
\name{get_sensitivity}
\alias{get_sensitivity}
\title{This function examines how the optimal installed capacities of an ANTARES study will be impacted by solutions in the neighborhood of the optimal cost.}
\usage{
get_sensitivity(epsilon = 0, opts = antaresRead::simOptions())
}
\arguments{
\item{epsilon}{Epsilon (need to be taller than the optimality gap)}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{display}{Logical. If \code{TRUE} the advancement of the function
if displayed in the console}
}
\value{

}
\description{
- Launch this function after you launched benders Xpansion a first time
  - The range given for each candidate means that alternate solution of installed capacity are include in this range.
}
