% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{get_whole_simulation_name}
\alias{get_whole_simulation_name}
\title{Get the whole simulation name of an ANTARES simulation,
as it appears in the output path,containing also the 
date and hour identifier.}
\usage{
get_whole_simulation_name(name, opts = simOptions())
}
\arguments{
\item{name}{Name of the simulation}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function returns the most recent simulation name which contains
the argument \code{name}. The returned simulation name also includes the
identifier which are automatically added by ANTARES in the output path :
yyyymmdd-hhmmxxx-name.
}
\description{
Get the whole simulation name of an ANTARES simulation,
as it appears in the output path,containing also the 
date and hour identifier.
}

