% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{initiate_master}
\alias{initiate_master}
\title{Initiate master problem}
\usage{
initiate_master(candidates, exp_options, opts = antaresRead::simOptions())
}
\arguments{
\item{candidates}{list of investment candidates, as returned by
\code{\link{read_candidates}}}

\item{exp_options}{list of options related to the expansion planning, as returned
by the function \code{\link{read_options}}}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
This function does not return anything.
}
\description{
\code{initiate_master} copy the AMPL files into the temporary 
folder of the expansion planning optimisation and create the other input
and output file of the master problem
}
