% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_year_investment.R
\name{multi_year_investment}
\alias{multi_year_investment}
\title{Multi-year investment optimization}
\usage{
multi_year_investment(path_solver, directory_path = getwd(), display = TRUE,
  report = TRUE, clean = TRUE, parallel = TRUE)
}
\arguments{
\item{path_solver}{Character containing the Antares Solver path}

\item{directory_path}{Character containing the directory path in which are stored the studies.ini, candidates.ini and settings.ini files}

\item{display}{Logical. If \code{TRUE} the advancement of the optimisation algorithm
if displayed in the console}

\item{report}{Logical. If \code{TRUE} an html report of the expansion results will
be generated.}

\item{clean}{Logical. If \code{TRUE} the output of the ANTARES simulations run by the
package will be deleted (except for the output of the simulation which brings
to the best solution).}

\item{parallel}{Logical. If \code{TRUE} the ANTARES simulations will be run in parallel mode (Work
only with ANTARES v6.0.0 or more). In that case, the number of cores used by the simulation
is the one set in advanced_settings/simulation_cores (see ANTARES interface).}
}
\value{

}
\description{
Multi-year investment optimization
}
