% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_numeric_accuracy.R
\name{numeric_patch_cut}
\alias{numeric_patch_cut}
\title{This function update the bender cuts so as to deal with the numeric
inaccuracy of the ANTARES output
Not used since v0.3}
\usage{
numeric_patch_cut(x_current, x_max)
}
\arguments{
\item{x_current}{vector of current invested capacities}

\item{x_max}{vector of maximum possible investments}
}
\value{
Return correct term to substract to the b of \code{ax+b}
}
\description{
Marginal costs are returned in ANTARES with a 10^0 accuracy (without any decimal).
This can cause numeric imprecisions in the cut of the benders decomposition and
make it converge to a non-optimal solution.
}
\details{
This function (tries to) corrects this by updating the constant term of the cuts
}

