% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_options.R
\name{read_options}
\alias{read_options}
\title{Read expansion planning options}
\usage{
read_options(file, opts = antaresRead::simOptions())
}
\arguments{
\item{file}{path toward the text file which contains the settings}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Returns a list containing the different optimisation options
}
\description{
\code{read_options} is a function which read the options related to the
expansion planning optimization. The options are usually stored in the file 
antaresStudyPath/user/expansion/settings.ini.
}

