% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_studies.R
\name{read_studies}
\alias{read_studies}
\title{Read studies.ini file}
\usage{
read_studies(studies_file_name)
}
\arguments{
\item{studies_file_name}{character. file path of a studies.ini file}
}
\value{
Returns a list containing the different paths of the studies
}
\description{
The studies.ini file contains path towards several Antares studies indexed
by the year their represent. For instance \code{2025 = D:/antares1}.
}
