% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Write report which summarizes the results of the expansion
planning}
\usage{
report(benders_out = "last", file = default_report_file(opts),
  opts = antaresRead::simOptions())
}
\arguments{
\item{benders_out}{Output file of the benders decomposition function, as returned by
\code{\link{benders}}. If default value \code{last} is used, the function
will use the results of the last launched benders decomposition}

\item{file}{File name of the report. By default the report will be stored in the
directory user/expansion/report of the current ANTARES study}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{

}
\description{
Write report which summarizes the results of the expansion
planning
}

