% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_years.R
\name{select_years}
\alias{select_years}
\title{Identify some representative time series of an ANTARES project containing a significant amount of Monte-Carlo years}
\usage{
select_years(mainAreas = "fr", extraAreas = NULL, nMCyears = 5,
  weightMain = 0.5, weightExtra = 0, weightPeakMain = 0.5,
  weightPeakExtra = 0, subtractUnavoidableEnergyMain = TRUE,
  subtractUnavoidableEnergyExtra = TRUE,
  subtractNuclearAvailabilityMain = TRUE,
  subtractNuclearAvailabilityExtra = FALSE, showCurves = TRUE,
  showTable = TRUE, opts = antaresRead::simOptions())
}
\arguments{
\item{mainAreas}{Vector containing the names of the areas on which the clustering algorithm will be based.}

\item{extraAreas}{Vector containing the names of the additional areas on which the clustering algorithm will be based.
Usually it contains areas that need to be taken into accounts in the algorithm but in a smaller scale than the ones in mainAreas.
If \code{NULL}, no extra area is imported.}

\item{nMCyears}{Numeric representing the amount of Monte-Carlo years to get after the function.}

\item{weightMain}{Numeric (between 0 and 1) giving the weighting of the load monotonous for the main areas into the clustering algorithm choices.
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}

\item{weightExtra}{Numeric (between 0 and 1) giving the weighting of the load monotonous for the additional areas into the clustering algorithm choices.
It is usually lower than weightMain. 
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}

\item{weightPeakExtra}{Numeric (between 0 and 1) giving the weighting of the peak period (20 most crucial hours) for the additional areas into the clustering algorithm choices. 
It is usually lower than weightPeakMain. 
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}

\item{subtractUnavoidableEnergyMain}{If \code{TRUE}, unavoidable energy (Solar, Wind, etc.) is subtracted from LOAD of the main areas.}

\item{subtractUnavoidableEnergyExtra}{If \code{TRUE}, unavoidable energy (Solar, Wind, etc.) is subtracted from LOAD of the additional areas.}

\item{subtractNuclearAvailabilityMain}{If \code{TRUE}, nuclear availability is subtracted from LOAD of the main areas.}

\item{subtractNuclearAvailabilityExtra}{If \code{TRUE}, nuclear availability is subtracted from LOAD of the additional areas.}

\item{showCurves}{If \code{TRUE}, the function displays a load monotonous curve analysis}

\item{showTable}{If \code{TRUE}, the function displays a cost analysis (OP. COST, LOLD, UNSP. ENRG)
@param opts
List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{weightPeakMean}{Numeric (between 0 and 1) giving the weighting of the peak period (20 most crucial hours) for the main areas into the clustering algorithm choices. 
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}
}
\value{

}
\description{
Identify some representative time series of an ANTARES project containing a significant amount of Monte-Carlo years
}
