% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benders_tools.R
\name{set_antares_options}
\alias{set_antares_options}
\title{Set ANTARES study options related to the current benders decomposition}
\usage{
set_antares_options(benders_options, candidates,
  opts = antaresRead::simOptions())
}
\arguments{
\item{benders_options}{list of benders decomposition options, as returned by
\code{\link{read_options}}.}

\item{candidates}{list of investment candidates, as returned by
\code{\link{read_candidates}}}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
nothing
}
\description{
Set ANTARES study options related to the current benders decomposition
}
