% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{set_playlist}
\alias{set_playlist}
\title{Set playlist of the study}
\usage{
set_playlist(playlist, opts = simOptions())
}
\arguments{
\item{playlist}{vector of MC years identifier to be simulated}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{set_playlist} is a function which modifies the input file of an ANTARES
study and set the playlist in order to simulate only the given MC years
}

