% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{set_simulation_period}
\alias{set_simulation_period}
\title{Set period on which we want the ANTARES simulation to run.
Update "first day" and "last day" parameters of the ANTARES study.}
\usage{
set_simulation_period(weeks, opts = antaresRead::simOptions())
}
\arguments{
\item{weeks}{vector of weeks identiers which should be simulated}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{

}
\description{
Set period on which we want the ANTARES simulation to run.
Update "first day" and "last day" parameters of the ANTARES study.
}

