% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_settings.R
\name{set_uc_mode}
\alias{set_uc_mode}
\title{Set the unit-commitment mode of the ANTARES study}
\usage{
set_uc_mode(mode, opts = antaresRead::simOptions())
}
\arguments{
\item{mode}{Unit-commitment mode. Must be equal to "fast", "accurate"}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{set_uc_mode} is a function which modifies the input file of an ANTARES
study and set the simulation setting "unit-commitment-mode"
}
