% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.r
\name{to}
\alias{to}
\title{Return the second area that a link connects}
\usage{
to(link_name)
}
\arguments{
\item{link_name}{Name of one link}
}
\value{
name of the second area that the link connects
}
\description{
Return the second area that a link connects
}

