% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benders_tools.R
\name{update_average_cuts}
\alias{update_average_cuts}
\title{Update costs and rentability files of the average cuts
Create files in_avgrentability.txt and in_avgcuts.txt which
will be later read by AMPL as input of the master problem}
\usage{
update_average_cuts(current_it, candidates, output_link_s, output_link_h_s,
  ov_cost, n_w, tmp_folder, benders_options)
}
\arguments{
\item{current_it}{list of current iteration characteristics}

\item{candidates}{list of investment candidates, as returned by
\code{\link{read_candidates}}}

\item{output_link_s}{antaresDataList of all the links of the study with a yearly time step
and averaged on all MC years}

\item{output_link_h_s}{antaresDataList of all the links of the study with an hourly time step
and averaged on all MC years}

\item{ov_cost}{overall costs of this iteration}

\item{n_w}{number of weeks per year}

\item{tmp_folder}{temporary folder in which to write the files}

\item{benders_options}{list of benders decomposition options, as returned by
\code{\link{read_options}}.}
}
\value{
nothing
}
\description{
Update costs and rentability files of the average cuts
Create files in_avgrentability.txt and in_avgcuts.txt which
will be later read by AMPL as input of the master problem
}

