% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.r
\name{update_link}
\alias{update_link}
\title{Update the properties of a link (capacity, impedance or hurdle costs)}
\usage{
update_link(link_name, property_name, new_value,
  opts = antaresRead::simOptions())
}
\arguments{
\item{link_name}{Name of one link}

\item{property_name}{Name of the propery to change, must be equal to "direct_capacity" or "indirect_capacity" or
"impedance" or "direct_hurdle_cost" or "indirect_hurdle_cost"}

\item{new_value}{New value of the property}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
ANTARES study
}
\description{
\code{update_link} is a function which modifies the input file of an ANTARES
study and update the properties of a link
}

