% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benders_tools.R
\name{update_weekly_cuts}
\alias{update_weekly_cuts}
\title{Update costs and rentability files of the weekly cuts
Create files in_weeklyrentability.txt and in_weeklycuts.txt which
will be later read by AMPL as input of the master problem}
\usage{
update_weekly_cuts(current_it, candidates, output_area_w, output_link_w,
  output_link_h, inv_cost, tmp_folder, benders_options)
}
\arguments{
\item{current_it}{list of current iteration characteristics}

\item{candidates}{list of investment candidates, as returned by
\code{\link{read_candidates}}}

\item{output_area_w}{antaresDataList of all the areas of the study with a weekly time step}

\item{output_link_w}{antaresDataList of all the links of the study with a weekly time step}

\item{output_link_h}{antaresDataList of all the links of the study with a weekly time step}

\item{inv_cost}{investments costs of this iteration}

\item{tmp_folder}{temporary folder in which to write the files}

\item{benders_options}{list of benders decomposition options, as returned by
\code{\link{read_options}}.}
}
\value{
nothing
}
\description{
Update costs and rentability files of the weekly cuts
Create files in_weeklyrentability.txt and in_weeklycuts.txt which
will be later read by AMPL as input of the master problem
}

