% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benders_tools.R
\name{update_yearly_cuts}
\alias{update_yearly_cuts}
\title{Update costs and rentability files of the yearly cuts
Create files in_yearlyrentability.txt and in_yearlycuts.txt which
will be later read by AMPL as input of the master problem}
\usage{
update_yearly_cuts(current_it, candidates, output_area_y, output_link_y,
  inv_cost, n_w, tmp_folder)
}
\arguments{
\item{current_it}{list of current iteration characteristics}

\item{candidates}{list of investment candidates, as returned by
\code{\link{read_candidates}}}

\item{output_area_y}{antaresDataList of all the areas of the study with a yearly time step}

\item{output_link_y}{antaresDataList of all the links of the study with a yearly time step}

\item{inv_cost}{investments costs of this iteration}

\item{n_w}{number of weeks per year}

\item{tmp_folder}{temporary folder in which to write the files}
}
\value{
nothing
}
\description{
Update costs and rentability files of the yearly cuts
Create files in_yearlyrentability.txt and in_yearlycuts.txt which
will be later read by AMPL as input of the master problem
}

