% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_selection.R
\name{week_selection}
\alias{week_selection}
\title{Smart selection of the weeks to simuate}
\usage{
week_selection(current_it, mc_years, weeks, tmp_folder, exp_options)
}
\arguments{
\item{current_it}{list of current iteration characteristics}

\item{mc_years}{vector of all the Monte Carlo years identifier}

\item{weeks}{vector of all the weeks identifier}

\item{tmp_folder}{temporary folder of the benders decomposition}

\item{exp_options}{list of options related to the expansion planning, as returned
by the function \code{\link{read_options}}}
}
\value{
updated current iteration characteristics
}
\description{
week_selection is a function which select the MC years and 
weeks which will be simulated during this iteration
}
