% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_candidates.R
\name{with_profile}
\alias{with_profile}
\title{Return vector of candidate names which have a link profile}
\usage{
with_profile(candidates)
}
\arguments{
\item{candidates}{list of investment candidates, as returned by
\code{\link{read_candidates}}}
}
\value{
Returns a vector of link name
}
\description{
Return vector of candidate names which have a link profile
}
