% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benders.R
\name{benders}
\alias{benders}
\title{Launch benders decomposition}
\usage{
benders(path_solver, display = TRUE, report = TRUE, clean = TRUE,
  parallel = TRUE, recovery_mode = FALSE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{path_solver}{Character containing the Antares Solver path}

\item{display}{Logical. If \code{TRUE} the advancement of the benders decomposition
if displayed in the console}

\item{report}{Logical. If \code{TRUE} an html report of the expansion results will
be generated.}

\item{clean}{Logical. If \code{TRUE} the output of the ANTARES simulations run by the
package will be deleted (except for the output of the simulation which brings
to the best solution).}

\item{parallel}{Logical. If \code{TRUE} the ANTARES simulations will be run in parallel mode (Work
only with ANTARES v6.0.0 or more). In that case, the number of cores used by the simulation
is the one set in advanced_settings/simulation_cores (see ANTARES interface).}

\item{recovery_mode}{Logical. If \code{TRUE} will launch the benders decomposition keeping the cut files already saved
in the temporary folder of the ANTARES study. Can be used if a previous simulation has crashed.}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{

}
\description{
Launch benders decomposition
}

