% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_candidates.R
\name{read_candidates}
\alias{read_candidates}
\title{Read investment candidates with their characteristics}
\usage{
read_candidates(opts = antaresRead::simOptions())
}
\arguments{
\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Returns a list containing the different investment candidates.
}
\description{
\code{read_candidates} is a function which read the investments candidates
of the expansion planning problem and their characteristics. The information on
the candidates is stored in the file antaresStudyPath/user/expansion/candidates.ini.
}

