% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_years.R
\name{select_years}
\alias{select_years}
\title{Identify some representative time series of an ANTARES project}
\usage{
select_years(mainAreas = "fr", extraAreas = NULL, selection = 5,
  MCYears = "all", weightMain = 0.5, weightPeakMain = 0.5,
  weightExtra = 0, weightPeakExtra = 0,
  subtractUnavoidableEnergyMain = TRUE,
  subtractUnavoidableEnergyExtra = TRUE,
  subtractNuclearAvailabilityMain = TRUE,
  subtractNuclearAvailabilityExtra = FALSE, displayCurves = TRUE,
  displayTable = TRUE, opts = antaresRead::simOptions())
}
\arguments{
\item{mainAreas}{Vector containing the names of the areas on which the clustering algorithm will be based.
Usually \code{mainAreas} = \code{"fr"}.}

\item{extraAreas}{Vector containing the names of the additional areas on which the clustering algorithm will be based.
Usually it contains areas that need to be taken into account in the algorithm but in a smaller scale than the ones in \code{mainAreas}.
If \code{mainAreas} = \code{"fr"}, \code{extraAreas} must not contain \code{"fr"} again to be relevant enough.
If \code{NULL}, no extra area is imported.}

\item{selection}{Numeric representing the amount of Monte-Carlo years to get after the function.}

\item{MCYears}{Index of the Monte-Carlo years to import. 
If \code{"all"}, every MC years are read, else the specified Monte-Carlo simulations are imported.}

\item{weightMain}{Numeric (between 0 and 1) giving the weighting of the load monotonous for the main areas into the clustering algorithm choices.
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}

\item{weightExtra}{Numeric (between 0 and 1) giving the weighting of the load monotonous for the additional areas into the clustering algorithm choices.
It is usually lower than \code{WeightExtra}. 
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}

\item{weightPeakExtra}{Numeric (between 0 and 1) giving the weighting of the peak period (20 most crucial hours) for the additional areas into the clustering algorithm choices. 
It is usually lower than \code{WeightPeakMain}. 
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}

\item{subtractUnavoidableEnergyMain}{If \code{TRUE}, unavoidable energy (Solar, Wind, etc.) is subtracted from LOAD of the main areas.}

\item{subtractUnavoidableEnergyExtra}{If \code{TRUE}, unavoidable energy (Solar, Wind, etc.) is subtracted from LOAD of the additional areas.}

\item{subtractNuclearAvailabilityMain}{If \code{TRUE}, nuclear availability is subtracted from LOAD of the main areas.}

\item{subtractNuclearAvailabilityExtra}{If \code{TRUE}, nuclear availability is subtracted from LOAD of the additional areas.}

\item{displayCurves}{If \code{TRUE}, the function displays a load monotonous curve analysis}

\item{displayTable}{If \code{TRUE}, the function displays a cost analysis (OP. COST, LOLD, UNSP. ENRG)}

\item{opts}{List of simulation parameters returned by the function \code{antaresRead::setSimulationPath}}

\item{weightPeakMean}{Numeric (between 0 and 1) giving the weighting of the peak period (20 most crucial hours) for the main areas into the clustering algorithm choices. 
If \code{0}, no importance is given to this criteria. If \code{1}, the algorithm will be based only on this criteria.}
}
\value{
If \code{displayCurves} and \code{displayTable} are both \code{FALSE}, only identities and weightings of the selected Monte-Carlo years are displayed.
Else, curves of load monotonous for imported areas will be also plotted (for every MC years, clusters and average curves) and a cost analysis will be given (based on OP. COST, LOLD and UNSP. ENRG).
}
\description{
Identify some representative time series of an ANTARES project containing a significant amount of Monte-Carlo years.
Usually based on France (= \code{mainAreas}) load monotonous.
Europe can also be imported as \code{extraAreas} to take energy imports and exports into account.
}
\details{
When \code{subtractUnavoidableEnergyMain}, \code{subtractUnavoidableEnergyExtra}, \code{subtractNuclearAvailabilityMain} and \code{subtractNuclearAvailabilityExtra} are all \code{TRUE}, the function may crash because of insufficient memory. 
In such a case, it is necessary to reduce the size of the input.
Different strategies are available depending on your objective : reduce the number of Monte-Carlo years, do not subtract nuclear availability (especially on extra areas), take fewer areas, increase the memory (with setRam), etc.
}
\examples{
# Import simulation
setSimulationPath()

# Find 5 Monte-Carlo year clusters for the simulation
# Study France as the main area and europe as the secondary area
# Base algorithm on :
# 40\% for the load monotonous in France
# 40\% for the peak period in France
# 20\% for the load monotonous in all europe
# 0\% for the peak period in all europe
europe <- c("at", "be", "ch", "de", "es", "gb", "ie", "it", "lu_be", "lu_de", "ni", "nl", "pt")
select_years(mainAreas = "fr", extraAreas = europe, weightMain = 0.4, weightPeakMain = 0.4, weightExtra = 0.2)

}
